CxxTest testcase for week1 excercise 2.

+++ week1_programming_challenges/2_maximum_pairwise_product/max_pairwise_product.h
@@ -1,5 +1,6 @@
 #include <vector>
+#include <stdint.h>
 
 using std::vector;
 
-long long BetterMaxPairwiseProduct(const vector<long long>& numbers);
+uint64_t BetterMaxPairwiseProduct(const vector<uint64_t>& numbers);
